NNHanman: move filter classes to Filters.kt
